+++ admin/sections/config/model.tsx
@@ -28,9 +28,9 @@ import { useForm } from "react-hook-form";
 import { useMutation, useQuery } from "react-query";
 import { AzureOpenAIForm } from "./providers/azureOpenai";
 import { GeminiForm } from "./providers/gemini";
-import { GroqForm } from "./providers/groq";
 import { OllamaForm } from "./providers/ollama";
 import { OpenAIForm } from "./providers/openai";
+import { TSystemsForm } from "./providers/t-systems";
 
 export const ModelConfig = ({
   sectionTitle,
@@ -98,14 +98,14 @@ export const ModelConfig = ({
     switch (defaultValues.model_provider ?? "") {
       case "openai":
         return <OpenAIForm form={form} defaultValues={defaultValues} />;
-      case "groq":
-        return <GroqForm form={form} defaultValues={defaultValues} />;
       case "ollama":
         return <OllamaForm form={form} defaultValues={defaultValues} />;
       case "gemini":
         return <GeminiForm form={form} defaultValues={defaultValues} />;
       case "azure-openai":
         return <AzureOpenAIForm form={form} defaultValues={defaultValues} />;
+      case "t-systems":
+        return <TSystemsForm form={form} defaultValues={defaultValues} />;
       default:
         return null;
     }