+++ src/models/model_config.py
@@ -10,6 +10,11 @@ class OpenAIConfig(BaseModel):
         description="The OpenAI API key to use",
         env="OPENAI_API_KEY",
     )
+    openai_api_base: str | None = Field(
+        default=None,
+        description="The base URL for the OpenAI API",
+        env="OPENAI_API_BASE",
+    )
 
 
 class GeminiConfig(BaseModel):
@@ -17,14 +22,6 @@ class GeminiConfig(BaseModel):
         default=None,
         description="The Google API key to use",
         env="GOOGLE_API_KEY",
-    )
-
-
-class GroqConfig(BaseModel):
-    groq_api_key: str | None = Field(
-        default=None,
-        description="The Groq API key to use",
-        env="GROQ_API_KEY",
     )
 
 
@@ -69,15 +66,28 @@ class AzureOpenAIConfig(BaseModel):
     )
 
 
+class TSystemsConfig(BaseModel):
+    t_systems_llmhub_api_key: str | None = Field(
+        default=None,
+        description="The T-Systems LLMHub API key to use",
+        env="T_SYSTEMS_LLMHUB_API_KEY",
+    )
+    t_systems_llmhub_api_base: str | None = Field(
+        default="https://llm-server.llmhub.t-systems.net/v2",
+        description="The base URL for the T-Systems LLMHub API",
+        env="T_SYSTEMS_LLMHUB_BASE_URL",
+    )
+
+
 # We're using inheritance to flatten all the fields into a single class
 # Todo: Refactor API to nested structure
 class ModelConfig(
     BaseEnvConfig,
     OpenAIConfig,
-    GroqConfig,
     GeminiConfig,
     OllamaConfig,
     AzureOpenAIConfig,
+    TSystemsConfig,
 ):
     model_provider: str | None = Field(
         default=None,
@@ -117,8 +127,8 @@ class ModelConfig(
             return True
         elif self.model_provider == "azure-openai":
             return True
-        elif self.model_provider == "groq":
-            return self.groq_api_key is not None
+        elif self.model_provider == "t-systems":
+            return self.t_systems_llmhub_api_key is not None
         return False
 
     @classmethod